+++ src/app/mypage/(component)/WhMypageNavMenu.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import Link from 'next/link';
 import { useSelectedLayoutSegment } from 'next/navigation';
