+++ src/components/WhMoreButton.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import {
   ChevronDownIcon,
   ChevronUpIcon,
-} from '../../../../public/assets/icons/arrow';
+} from '../../public/assets/icons/arrow';
 
 const defaultMoreBtnCss =
   'absolute mt-2 max-w-[118px] z-10 flex gap-1 items-center py-2 px-3 border rounded-[20px] border-nutral-white-03 bg-nutral-white-01 text-nutral-black-03 text-caption-01';
